Add tests on creation of PR - CI stage (#8)
* Add tests on creation of PR - CI stage
Fixes #7

* updated with running FastAPI application

* Updating changes to Test API step

* Update CI Tests workflow

This commit updates the CI Tests workflow in the `.github/workflows/ci_tests.yml` file. The changes include fixing indentation, adding a step to wait for the API to be available, and listing running containers.

* Remove wait for API step in CI workflow

* Update API endpoint in CI test workflow and Docker Compose

This commit updates the API endpoint in the CI test workflow to use the "api" hostname instead of "localhost". It also updates the Docker Compose file to use the "fastapi-app" image instead of "testappfastapi".

* Update API URL in CI test script

---------

Co-authored-by: Amruth VVKP <amruthvvkp@gmail.com>